// cliente-movil/front/src/app/components/report-market-form/report-market-form.types.ts
export interface MarketOptionDTO{
    id: string;
    name: string;
}

export interface ProductOptionDTO{
    id: string;
    name: string;
}

export interface MarketDTO {
    id: string;
    name: string;
    address: string;
    urlImg: string;
}

export interface ProductDTO {
    id: string;
    name: string;
    urlImg: string;
    detail: string;
}

export interface MarketReportForm {
    marketSelected: MarketDTO;
    productSelected: ProductDTO;
    reportContent: string;
  }